requirements.txt
streamlit
torch
opencv-python
numpy
pandas
Pillow
pydrive
firebase-admin
